+++ workflows/core/steps/doc.ts
@@ -9,7 +9,7 @@ import { contextFromInput } from "../utils.ts";
 import { raise } from "xstate";
 import path from "node:path";
 import { getSourceUrl } from "../store.ts";
-import { handlePrompt } from "../xstate-actions/prompt.ts";
+import { handlePrompt } from "../prompt.ts";
 
 /**
  * Input for the DocStepMachine.
